AI-processing Services/real_estate_chatbot/requirements.txt
flask
psycopg2
python-dotenv
ollama
numpy
pyodbc